updata

+++ THUC_HANH_TREN_LOP/Buoi8/thuchanh1/Employee.java
@@ -9,7 +9,7 @@ public class Employee {
     {
         return  this.hours;
     }
-    public int getSalary()
+    public double getSalary()
     {
         return this.salary;
     }
